Yet another retype for MSVC/qt6

+++ src/map/IMG/rgnfile.cpp
@@ -14,8 +14,8 @@ static quint64 pointId(const QPoint &pos, quint32 type, quint32 labelPtr)
 {
 	quint64 id;
 
-	uint hash = (uint)qHash(QPair<uint,uint>(qHash(QPair<int, int>(pos.x(),
-	  pos.y())), labelPtr & 0x3FFFFF));
+	uint hash = (uint)qHash(QPair<uint,uint>((uint)qHash(
+	  QPair<int, int>(pos.x(), pos.y())), labelPtr & 0x3FFFFF));
 	id = ((quint64)type)<<32 | hash;
 	// Make country labels precedent over city labels
 	if (!(type >= 0x1400 && type <= 0x153f))
